[TASK] Raise minimum PHP constraint in composer.json
This package uses php functionality only available
in PHP 8.0 and some only in PHP8.1. This is a hard
requirement, as it needs to scan and handle code
for these versions using the `Reflection` API.

Thus this change raises the minimum PHP version to
ensure it is run with the correct supported PHP
versions.

Uses command(s):

> composer req --no-update php:"^8.1 || ^8.2"